Bump rustc version to 1.66.0

+++ misc/src/quick_sort.rs
@@ -80,7 +80,7 @@ fn quick_sort_in_place<T: Ord>(arr: &mut [T]) {
                 arr.swap(i as usize, j as usize);
             }
         }
-        arr.swap(i as usize, pivot as usize);
+        arr.swap(i as usize, pivot);
         i
     }
 }
